refactor: 패키지명 및 클래스명 의미에 맞게 변경
