+++ subtysis/src/main/java/com/hackday/subtysis/metadatatype/BaseMetadataType.java
@@ -1,10 +1,10 @@
-package com.hackday.subtysis.metadata;
+package com.hackday.subtysis.metadatatype;
 
 import java.lang.reflect.Type;
 
 /**
  * Create by Yeji on 22,June,2020.
  */
-public abstract class BaseMetadata {
+public abstract class BaseMetadataType {
     public abstract Type getListType();
 }
