+++ subtysis/src/main/java/com/hackday/subtysis/metadatatype/EncyclopediaMetadataType.java
@@ -1,4 +1,4 @@
-package com.hackday.subtysis.metadata;
+package com.hackday.subtysis.metadatatype;
 
 import com.google.gson.reflect.TypeToken;
 import com.hackday.subtysis.model.items.EncyclopediaItem;
@@ -9,7 +9,7 @@ import java.util.ArrayList;
 /**
  * Create by Yeji on 22,June,2020.
  */
-public class EncyclopediaMetadata extends BaseMetadata {
+class EncyclopediaMetadataType extends BaseMetadataType {
     @Override
     public Type getListType() {
         return new TypeToken<ArrayList<EncyclopediaItem>>() {}.getType();
