+++ subtysis/src/main/java/com/hackday/subtysis/metadatatype/MetadataTypeFactory.java
@@ -1,4 +1,4 @@
-package com.hackday.subtysis.metadata;
+package com.hackday.subtysis.metadatatype;
 
 import com.hackday.subtysis.model.SearchType;
 
@@ -6,5 +6,5 @@ import com.hackday.subtysis.model.SearchType;
  * Create by Yeji on 22,June,2020.
  */
 public interface MetadataTypeFactory {
-    BaseMetadata createInstance(SearchType searchType);
+    BaseMetadataType createInstance(SearchType searchType);
 }
